refactor: move SDK metadata extensions to a separate file
