+++ lib/bloc/trezor_init_bloc/trezor_init_bloc.dart
@@ -14,6 +14,7 @@ import 'package:web_dex/model/hw_wallet/init_trezor.dart';
 import 'package:web_dex/model/hw_wallet/trezor_status.dart';
 import 'package:web_dex/model/hw_wallet/trezor_status_error.dart';
 import 'package:web_dex/model/hw_wallet/trezor_task.dart';
+import 'package:web_dex/model/kdf_auth_metadata_extension.dart';
 import 'package:web_dex/model/text_error.dart';
 import 'package:web_dex/model/wallet.dart';
 import 'package:web_dex/shared/utils/utils.dart';
